pyproject.toml
[project]
name = "mise-a-niveau"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.120.1",
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.3",
    "ollama>=0.4.8",
    "pyyaml>=6.0.2",
]
